requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
azure-cosmos
pymongo
h5py
numpy
pandas
scikit-learn
matplotlib
seaborn
joblib
motor